+++ src/data/state.js
@@ -1,7 +1,7 @@
 import EutopicCore from '@ticketland-io/eutopic-core'
 import EutopicSolanaWallet from '@ticketland-io/eutopic-solana-wallet'
 import Enclave from '@ticketland-io/eutopic-web-enclave'
-import FirebaseAuth from '../data/record';
+import FirebaseAuth from './firebase';
 
 const Wallet = () => EutopicSolanaWallet({enclave: Enclave()})
 const eutopicCore = EutopicCore({Wallet})
